[#93[ style: 디테일 info 부분 다크모드 적용
