+++ client/components/ProductDetail/DetailBtns /DetailBtns.tsx
@@ -1,5 +1,6 @@
 import { ReactElement, useState } from 'react';
 import styled from '@lib/styledComponent';
+import * as S from './style';
 
 import HeartIcon from '@image/heartIcon.svg';
 
@@ -15,59 +16,12 @@ export default function DetailBtns({ id, isLiked }: Props): ReactElement {
   const handleLikeClick = () => setIsHeart((isHeart) => !isHeart);
 
   return (
-    <StyledDetailBtns isHeart={isHeart}>
+    <S.DetailBtns isHeart={isHeart}>
       <button className="like-btn" onClick={handleLikeClick}>
         <HeartIcon />
       </button>
-      <button className="chart-btn">장바구니</button>
+      <button className="cart-btn">장바구니</button>
       <button className="purchase-btn">바로 구매</button>
-    </StyledDetailBtns>
+    </S.DetailBtns>
   );
 }
-
-interface StyledProps {
-  isHeart: boolean;
-}
-
-const StyledDetailBtns = styled.div<StyledProps>`
-  display: flex;
-  justify-content: flex-end;
-  gap: 8px;
-
-  background-color: var(--white);
-
-  & > button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 50px;
-    padding: 3px;
-  }
-
-  .like-btn {
-    background-color: var(--white);
-    border: 1px solid var(--gray4);
-    width: 50px;
-    border-radius: 5px;
-
-    & > svg {
-      stroke-width: ${({ isHeart }) => (isHeart ? '0px' : '1px')};
-      fill: var(${({ isHeart }) => (isHeart ? '--red' : '--white')});
-    }
-  }
-
-  .chart-btn {
-    width: 170px;
-    color: var(--black);
-    background-color: var(--white);
-    border: 1px solid var(--gray4);
-    border-radius: 5px;
-  }
-
-  .purchase-btn {
-    width: 230px;
-    color: var(--text-white);
-    background-color: var(--mint);
-    border-radius: 5px;
-  }
-`;
